+++ chatmaild/src/chatmaild/delete_inactive_users.py
@@ -2,32 +2,30 @@
 Remove inactive users
 """
 
+import os
 import shutil
 import sys
 import time
 
 from .config import read_config
-from .database import Database
-from .doveauth import iter_userdb_lastlogin_before
 
 
-def delete_inactive_users(db, config, CHUNK=100):
+def delete_inactive_users(config):
     cutoff_date = time.time() - config.delete_inactive_users_after * 86400
+    for addr in os.listdir(config.mailboxes_dir):
+        try:
+            user = config.get_user(addr)
+        except ValueError:
+            continue
 
-    old_users = iter_userdb_lastlogin_before(db, cutoff_date)
-    chunks = (old_users[i : i + CHUNK] for i in range(0, len(old_users), CHUNK))
-    for sublist in chunks:
-        for user in sublist:
-            user_mail_dir = config.get_user_maildir(user)
-            shutil.rmtree(user_mail_dir, ignore_errors=True)
-
-        with db.write_transaction() as conn:
-            for user in sublist:
-                conn.execute("DELETE FROM users WHERE addr = ?", (user,))
+        read_timestamp = user.get_last_login_timestamp()
+        if read_timestamp and read_timestamp < cutoff_date:
+            path = config.mailboxes_dir.joinpath(addr)
+            assert path == user.maildir
+            shutil.rmtree(path, ignore_errors=True)
 
 
 def main():
     (cfgpath,) = sys.argv[1:]
     config = read_config(cfgpath)
-    db = Database(config.passdb_path)
-    delete_inactive_users(db, config)
+    delete_inactive_users(config)