+++ chatmaild/src/chatmaild/doveauth.py
@@ -3,18 +3,12 @@ import json
 import logging
 import os
 import sys
-import time
-from pathlib import Path
 
 from .config import Config, read_config
-from .database import Database
 from .dictproxy import DictProxy
+from .migrate_db import migrate_from_db_to_maildir
 
 NOCREATE_FILE = "/etc/chatmail-nocreate"
-
-
-class UnknownCommand(ValueError):
-    """dictproxy handler received an unkown command"""
 
 
 def encrypt_password(password: str):
@@ -61,93 +55,6 @@ def is_allowed_to_create(config: Config, user, cleartext_password) -> bool:
     return True
 
 
-def get_user_data(db, config: Config, user):
-    if user == f"echo@{config.mail_domain}":
-        return dict(
-            home=str(config.get_user_maildir(user)),
-            uid="vmail",
-            gid="vmail",
-        )
-
-    with db.read_connection() as conn:
-        result = conn.get_user(user)
-    if result:
-        result["home"] = str(config.get_user_maildir(user))
-        result["uid"] = "vmail"
-        result["gid"] = "vmail"
-    return result
-
-
-def lookup_userdb(db, config: Config, user):
-    return get_user_data(db, config, user)
-
-
-def lookup_passdb(db, config: Config, user, cleartext_password, last_login=None):
-    if user == f"echo@{config.mail_domain}":
-        # Echobot writes password it wants to log in with into /run/echobot/password
-        try:
-            password = Path("/run/echobot/password").read_text()
-        except Exception:
-            logging.exception("Exception when trying to read /run/echobot/password")
-            return None
-
-        return dict(
-            home=str(config.get_user_maildir(user)),
-            uid="vmail",
-            gid="vmail",
-            password=encrypt_password(password),
-        )
-
-    if last_login is None:
-        last_login = time.time()
-    last_login = int(last_login)
-
-    with db.write_transaction() as conn:
-        userdata = conn.get_user(user)
-        if userdata:
-            # Update last login time.
-            conn.execute(
-                "UPDATE users SET last_login=? WHERE addr=?", (last_login, user)
-            )
-
-            userdata["home"] = str(config.get_user_maildir(user))
-            userdata["uid"] = "vmail"
-            userdata["gid"] = "vmail"
-            return userdata
-        if not is_allowed_to_create(config, user, cleartext_password):
-            return
-
-        encrypted_password = encrypt_password(cleartext_password)
-        q = """INSERT INTO users (addr, password, last_login)
-               VALUES (?, ?, ?)"""
-        conn.execute(q, (user, encrypted_password, last_login))
-        print(f"Created address: {user}", file=sys.stderr)
-        return dict(
-            home=str(config.get_user_maildir(user)),
-            uid="vmail",
-            gid="vmail",
-            password=encrypted_password,
-        )
-
-
-def iter_userdb(db) -> list:
-    """Get a list of all user addresses."""
-    with db.read_connection() as conn:
-        rows = conn.execute(
-            "SELECT addr from users",
-        ).fetchall()
-    return [x[0] for x in rows]
-
-
-def iter_userdb_lastlogin_before(db, cutoff_date):
-    """Get a list of users where last login was before cutoff_date."""
-    with db.read_connection() as conn:
-        rows = conn.execute(
-            "SELECT addr FROM users WHERE last_login < ?", (cutoff_date,)
-        ).fetchall()
-    return [x[0] for x in rows]
-
-
 def split_and_unescape(s):
     """Split strings using double quote as a separator and backslash as escape character
     into parts."""
@@ -175,9 +82,8 @@ def split_and_unescape(s):
 
 
 class AuthDictProxy(DictProxy):
-    def __init__(self, db, config):
+    def __init__(self, config):
         super().__init__()
-        self.db = db
         self.config = config
 
     def handle_lookup(self, parts):
@@ -189,14 +95,13 @@ class AuthDictProxy(DictProxy):
         args = list(split_and_unescape(args))
 
         config = self.config
-        db = self.db
         reply_command = "F"
         res = ""
         if namespace == "shared":
             if type == "userdb":
                 user = args[0]
                 if user.endswith(f"@{config.mail_domain}"):
-                    res = lookup_userdb(db, config, user)
+                    res = self.lookup_userdb(user)
                 if res:
                     reply_command = "O"
                 else:
@@ -204,7 +109,7 @@ class AuthDictProxy(DictProxy):
             elif type == "passdb":
                 user = args[1]
                 if user.endswith(f"@{config.mail_domain}"):
-                    res = lookup_passdb(db, config, user, cleartext_password=args[0])
+                    res = self.lookup_passdb(user, cleartext_password=args[0])
                 if res:
                     reply_command = "O"
                 else:
@@ -215,16 +120,37 @@ class AuthDictProxy(DictProxy):
     def handle_iterate(self, parts):
         # example: I0\t0\tshared/userdb/
         if parts[2] == "shared/userdb/":
-            db = self.db
-            result = "".join(f"Oshared/userdb/{user}\t\n" for user in iter_userdb(db))
+            result = "".join(
+                f"Oshared/userdb/{user}\t\n" for user in self.iter_userdb()
+            )
             return f"{result}\n"
+
+    def iter_userdb(self) -> list:
+        """Get a list of all user addresses."""
+        return [x for x in os.listdir(self.config.mailboxes_dir) if "@" in x]
+
+    def lookup_userdb(self, addr):
+        return self.config.get_user(addr).get_userdb_dict()
+
+    def lookup_passdb(self, addr, cleartext_password):
+        user = self.config.get_user(addr)
+        userdata = user.get_userdb_dict()
+        if userdata:
+            return userdata
+        if not is_allowed_to_create(self.config, addr, cleartext_password):
+            return
+
+        user.set_password(encrypt_password(cleartext_password))
+        print(f"Created address: {user}", file=sys.stderr)
+        return user.get_userdb_dict()
 
 
 def main():
     socket, cfgpath = sys.argv[1:]
     config = read_config(cfgpath)
-    db = Database(config.passdb_path)
 
-    dictproxy = AuthDictProxy(db=db, config=config)
+    migrate_from_db_to_maildir(config)
+
+    dictproxy = AuthDictProxy(config=config)
 
     dictproxy.serve_forever_from_socket(socket)