+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## untagged
 
+- Migrate and remove sqlite database in favor of password/lastlogin tracking 
+  in a user's maildir.  This removes the need for "passdb" setting in ini file
+  which was introduced through #351 below. 
+  ([#379](https://github.com/deltachat/chatmail/pull/379))
+
 - BREAKING: new required chatmail.ini values:
 
   mailboxes_dir = /home/vmail/mail/{mail_domain}